Replaced Integer.valueOf() with Long.valueOf(), updated version to 1.4.4

+++ api-client/src/main/java/com/desk/java/apiclient/model/Link.java
@@ -97,6 +97,6 @@ public class Link implements Serializable {
                 || !StringUtils.isDigitsOnly(segment)) {
             return 0L;
         }
-        return Integer.valueOf(segment);
+        return Long.valueOf(segment);
     }
 }
